# openapi/ledger_spec.yaml
openapi: 3.0.1
info:
  description: Ledger API
  title: Ledger API
  version: v1.0.0
servers:
- url: http://ledger_url
paths:
...
  /v1/transaction:
    get:
      operationId: search transactions
      parameters:
      - description: Set to '2' to return failed transaction states FAILED_REJECTED/FAILED_EXPIRED/FAILED_CANCELLED
          mapped to declined/timedout/cancelled external status respectively.Otherwise
          these transaction states will all be mapped to `failed` status
        in: query
        name: status_version
        schema:
          type: integer
          format: int32
          default: 2
      - description: Set to 'true' to search for transactions by exact reference.
          Otherwise reference is partially matched
        in: query
        name: exact_reference_match
        schema:
          type: boolean
          default: false
      - example: test@example.org
        in: query
        name: email
        schema:
          type: string
      - example: my-payment-reference
        in: query
        name: reference
        schema:
          type: string
      - example: J Doe
        in: query
        name: cardholder_name
        schema:
          type: string
      - example: 7890
        in: query
        name: last_digits_card_number
        schema:
          type: string
      - example: 123456
        in: query
        name: first_digits_card_number
        schema:
          type: string
      - description: Comma delimited payment states.
        example: "success,error"
        in: query
        name: payment_states
        schema:
          $ref: '#/components/schemas/CommaDelimitedSetParameter'
      - example: success
        in: query
        name: state
        schema:
          type: string
      - description: Comma delimited refund states.
        example: "success,error"
        in: query
        name: refund_states
        schema:
          $ref: '#/components/schemas/CommaDelimitedSetParameter'
      - description: Comma delimited card brands.
        example: "visa,mastercard"
        in: query
        name: card_brands
        schema:
          $ref: '#/components/schemas/CommaDelimitedSetParameter'
      - description: From date of transactions to be searched (this date is inclusive).
        example: 2015-08-14T12:35:00Z
        in: query
        name: from_date
        schema:
          type: string
      - description: To date of transactions to be searched (this date is inclusive).
        example: 2015-08-14T12:35:00Z
        in: query
        name: to_date
        schema:
          type: string
      - example: PAYMENT
        in: query
        name: transaction_type
        schema:
          type: string
          enum:
          - PAYMENT
          - REFUND
          - DISPUTE
      - example: po_fj893joishj12lndk
        in: query
        name: gateway_payout_id
        schema:
          type: string
      - description: From date of transactions settled date to be searched (this date
          is inclusive).
        example: 2015-08-14
        in: query
        name: from_settled_date
        schema:
          type: string
      - description: To date of transactions settled date to be searched (this date
          is inclusive).
        example: 2015-08-14
        in: query
        name: to_settled_date
        schema:
          type: string
      - example: metadata-value-1
        in: query
        name: metadata_value
        schema:
          type: string
      - example: 17ii98mg7f6si930tcjt48ldlc
        in: query
        name: agreement_id
        schema:
          type: string
      - example: 1
        in: query
        name: page
        schema:
          type: integer
          format: int64
      - example: 100
        in: query
        name: display_size
        schema:
          type: integer
          format: int64
          default: 500
      - description: Set to 'true' to limit the search counting the total number of
          transactions to 'limit_total_size' param
        example: true
        in: query
        name: limit_total
        schema:
          type: boolean
          default: false
      - example: 1000
        in: query
        name: limit_total_size
        schema:
          type: integer
          format: int64
          default: 10000
      - example: a14f0926-b44d-4160-8184-1b1f66e576ab
        in: query
        name: gateway_transaction_id
        schema:
          type: string
      - description: Comma delimited dispute states.
        example: "won,needs_response"
        in: query
        name: dispute_states
        schema:
          $ref: '#/components/schemas/CommaDelimitedSetParameter'
      - description: Set to true to list transactions for all accounts.
        in: query
        name: override_account_id_restriction
        schema:
          type: boolean
      - description: Comma delimited gateway account IDs. Required except when override_account_id_restriction=true
        example: "1,2"
        in: query
        name: account_id
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionSearchResponse'
          description: OK
        "400":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing required query parameters
        "504":
          description: Search query cancelled due to query timeout
      summary: Search transactions by query params. Same endpoint can be used to download
        CSV (with  Accept header="text/csv"). Refer to code for details
      tags:
      - Transactions
  /v1/transaction/gateway-transaction/{gatewayTransactionId}:
    get:
      operationId: findByGatewayTransactionId
      parameters:
      - description: Transaction ID from payment provider
        example: a14f0926-b44d-4160-8184-1b1f66e576ab
        in: path
        name: gatewayTransactionId
        required: true
        schema:
          type: string
      - example: sandbox
        in: query
        name: payment_provider
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/TransactionView'
          description: OK
        "404":
          description: Not found
        "422":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: If payment_provider query parameter is missing
      summary: Get transaction for a gateway transaction ID
      tags:
      - Transactions
  /v1/transaction/redact-reference/{transactionExternalId}:
    post:
      operationId: redactReference
      parameters:
      - in: path
        name: transactionExternalId
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json; qs=1: {}
          description: default response
      tags:
      - Transactions
  /v1/transaction/{parentTransactionExternalId}/transaction:
    get:
      operationId: getTransactionsForParentTransaction
      parameters:
      - description: Parent transaction external ID
        example: d0sk01d9amdk3ks0dk2dj03kd
        in: path
        name: parentTransactionExternalId
        required: true
        schema:
          type: string
      - description: Gateway account ID
        example: 1
        in: query
        name: gateway_account_id
        required: true
        schema:
          type: string
      - description: The type of transactions to return
        example: REFUND
        in: query
        name: transaction_type
        schema:
          type: string
          enum:
          - PAYMENT
          - REFUND
          - DISPUTE
      responses:
        "200":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/TransactionsForTransactionResponse'
          description: OK
        "404":
          description: Not found
        "422":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing required query parameter (gateway_account_id)
      summary: "Get transactions (ex: refunds) related to parent transaction (payment)"
      tags:
      - Transactions
  /v1/transaction/{transactionExternalId}:
    get:
      operationId: getById
      parameters:
      - in: path
        name: transactionExternalId
        required: true
        schema:
          type: string
      - in: query
        name: account_id
        schema:
          type: string
      - in: query
        name: override_account_id_restriction
        schema:
          type: boolean
      - in: query
        name: transaction_type
        schema:
          type: string
          enum:
          - PAYMENT
          - REFUND
          - DISPUTE
      - in: query
        name: parent_external_id
        schema:
          type: string
      - in: query
        name: status_version
        schema:
          type: integer
          format: int32
          default: 2
      responses:
        "200":
          content:
            application/json; qs=1:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionView'
          description: OK
        "400":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing required query parameters
        "500":
          description: Internal Server Error for invalid query parameter values
      summary: Get transaction by external ID
      tags:
      - Transactions
  /v1/transaction/{transactionExternalId}/event:
    get:
      operationId: events
      parameters:
      - description: Transaction external ID
        example: 9np5pocnotgkpp029d5kdfau5f
        in: path
        name: transactionExternalId
        required: true
        schema:
          type: string
      - example: 1
        in: query
        name: gateway_account_id
        required: true
        schema:
          type: string
      - description: Set to 'true' to return all events. By default events that do
          not map to an external state are removed and duplicate events mapping to
          same external state are reduced to one event.
        in: query
        name: include_all_events
        schema:
          type: boolean
          default: false
      - description: Set to '2' to return failed transaction states FAILED_REJECTED/FAILED_EXPIRED/FAILED_CANCELLED
          mapped to declined/timedout/cancelled status respectively.Otherwise these
          transaction states will all be mapped to `failed` status
        in: query
        name: status_version
        schema:
          type: integer
          format: int32
          default: 2
      responses:
        "200":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/TransactionEventResponse'
          description: OK
        "404":
          description: Not found
        "422":
          content:
            application/json; qs=1:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing required query parameters
      summary: Get events for transaction external ID
      tags:
      - Transactions
components:
  schemas:
...
    AuthorisationSummary:
      type: object
      properties:
        three_d_secure:
          $ref: '#/components/schemas/ThreeDSecure'
...
    CommaDelimitedSetParameter:
      type: object
      properties:
        notEmpty:
          type: boolean
        parameters:
          type: array
          items:
            type: string
          uniqueItems: true
        rawString:
          type: string
    ErrorResponse:
      type: object
      properties:
        error_identifier:
          type: string
          enum:
          - GENERIC
          - REFUND_NOT_AVAILABLE
          - REFUND_AMOUNT_AVAILABLE_MISMATCH
          - REFUND_NOT_AVAILABLE_DUE_TO_DISPUTE
          - ZERO_AMOUNT_NOT_ALLOWED
          - MOTO_NOT_ALLOWED
          - CANCEL_CHARGE_FAILURE_DUE_TO_CONFLICTING_TERMINAL_STATE_AT_GATEWAY_CHARGE_STATE_FORCIBLY_TRANSITIONED
          - CANCEL_CHARGE_FAILURE_DUE_TO_CONFLICTING_TERMINAL_STATE_AT_GATEWAY_INVALID_STATE_TRANSITION
          - AUTH_TOKEN_INVALID
          - AUTH_TOKEN_REVOKED
          - ACCOUNT_NOT_LINKED_WITH_PSP
          - TELEPHONE_PAYMENT_NOTIFICATIONS_NOT_ALLOWED
          - AGREEMENT_NOT_FOUND
          - ONE_TIME_TOKEN_INVALID
          - ONE_TIME_TOKEN_ALREADY_USED
          - INVALID_ATTRIBUTE_VALUE
          - CARD_NUMBER_REJECTED
          - AUTHORISATION_API_NOT_ALLOWED
          - AUTHORISATION_REJECTED
          - AUTHORISATION_ERROR
          - AUTHORISATION_TIMEOUT
          - AGREEMENT_NOT_ACTIVE
          - INCORRECT_AUTHORISATION_MODE_FOR_SAVE_PAYMENT_INSTRUMENT_TO_AGREEMENT
          - MISSING_MANDATORY_ATTRIBUTE
          - UNEXPECTED_ATTRIBUTE
          - ACCOUNT_DISABLED
          - RECURRING_CARD_PAYMENTS_NOT_ALLOWED
          - IDEMPOTENCY_KEY_USED
          - CARD_NUMBER_IN_PAYMENT_LINK_REFERENCE_REJECTED
          example: GENERIC
        message:
          type: array
          items:
            type: string
            example: example error message
...
    SettlementSummary:
      type: object
      properties:
        settled_date:
          type: string
          example: 2022-01-26
...
    TransactionEventResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TransactionEvent'
        transaction_id:
          type: string
          example: 9np5pocnotgkpp029d5kdfau5f
    TransactionSearchResponse:
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/PaginationBuilder'
        count:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/TransactionView'
        total:
          type: integer
          format: int64
...
    TransactionView:
      type: object
      properties:
        agreement_id:
          type: string
        amount:
          type: integer
          format: int64
          example: 1000
        authorisation_mode:
          type: string
          enum:
          - web
          - moto_api
          - agreement
          - external
          example: web
        authorisation_summary:
          $ref: '#/components/schemas/AuthorisationSummary'
        card_details:
          $ref: '#/components/schemas/CardDetails'
        corporate_card_surcharge:
          type: integer
          format: int64
          example: 0
        created_date:
          type: string
          format: date-time
        credential_external_id:
          type: string
          example: 7a7a718cb8b0401f8b8f4cdda9804073
        delayed_capture:
          type: boolean
        description:
          type: string
          example: New passport application
        disputed:
          type: boolean
        email:
          type: string
          example: citizen@example.org
        evidence_due_date:
          type: string
          format: date-time
        fee:
          type: integer
          format: int64
          example: 2
        gateway_account_id:
          type: string
          example: "1"
        gateway_payout_id:
          type: string
          example: po_mdoiu23jkdj1kj23sd
        gateway_transaction_id:
          type: string
          example: fd21-1mdknkls1-2121-csdf
        language:
          type: string
          example: en
        live:
          type: boolean
        metadata:
          type: object
          additionalProperties:
            type: object
        moto:
          type: boolean
        net_amount:
          type: integer
          format: int64
          example: -2
        parent_transaction_id:
          type: string
          description: Parent transaction external ID. Available for refunds
          example: 4dfdhapsdgftgkpp029d5kdfau5f
        payment_details:
          $ref: '#/components/schemas/TransactionView'
        payment_provider:
          type: string
          example: sandbox
        reason:
          type: string
          example: fraudulent
        reference:
          type: string
          example: payment reference
        refund_summary:
          $ref: '#/components/schemas/RefundSummary'
        refunded_by:
          type: string
          description: External ID of the service user that refunded payment. Only
            available for refunds
          example: dj8923eihsdih23hkjfe8d
        refunded_by_user_email:
          type: string
          description: Service user email that refunded the payment. Only available
            for refunds
          example: test@example.org
        return_url:
          type: string
          example: https://service-name.gov.uk/transactions/12345
        service_id:
          type: string
          example: baea0585e30d4acbb438524fca82b51e
        settlement_summary:
          $ref: '#/components/schemas/SettlementSummary'
        source:
          type: string
          enum:
          - CARD_API
          - CARD_PAYMENT_LINK
          - CARD_AGENT_INITIATED_MOTO
          - CARD_EXTERNAL_TELEPHONE
          example: CARD_API
        state:
          $ref: '#/components/schemas/ExternalTransactionState'
        total_amount:
          type: integer
          format: int64
          example: 1000
        transaction_id:
          type: string
          example: 9np5pocnotgkpp029d5kdfau5f
        transaction_type:
          type: string
          enum:
          - PAYMENT
          - REFUND
          - DISPUTE
          example: PAYMENT
        wallet_type:
          type: string
          example: APPLE_PAY
    TransactionsForTransactionResponse:
      type: object
      properties:
        parent_transaction_id:
          type: string
          example: 9np5pocnotgkpp029d5kdfau5f
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionView'
...